+++ apps/peer-rtc-dashboard/src/modules/settings/settings-appearance/index.ts
@@ -1,0 +1,1 @@
+export { SettingsAppearance } from '#/modules/settings/settings-appearance/settings-appearance';